No pending state calculation
- should have much less CPU load for non-mining clients
- --mine will not work
- For API calls, 'pending' will not work as parameter to API calls
- no warnings

The background task to create pending state is disabled, which
runs independently of the --mine flag to create a 'next block'
state to be served to respective API calls.

This takes all incoming transactions that are NOT in a block
yet, validates and runs them, to create the most current state
prediction possible, though for little use.

The mainline source is to be adapted in the future to also not
create this background load when it is not needed.